Text fills the available screen space with the chosen font size
The font size is selected based on the goal number of characters
we want to show per line. However, it is not always the case that
the desired width has a matching font size for the current resolution.

The smaller font size was selected to ensure that the desired amount
of text will fit. However, this means that some of the screen space
to the right was un-used.

This patch allows the line to fill the screen instead of being limited
to text->size. text->size is really just a suggestion, not a rule.